+++ include/down_quat.h
@@ -1,5 +1,5 @@
 #ifndef DOWN_QUAT_H
 
 #define DOWN_QUAT_H
-void down_quat(float* from, float* to, float* q_body_to_eci, float* goal_q);
+void down_quat(double* from, double* to, double* q_body_to_eci, double* goal_q);
 #endif
