+++ include/ecef2eci.h
@@ -1,4 +1,4 @@
 #ifndef ECEF2ECI
 #define ECEF2ECI
-void ecef_2_eci(float* ecef, float *eci, int unix);
+void ecef_2_eci(double* ecef, double* eci, int unix);
 #endif
