+++ include/kepler.h
@@ -1,4 +1,4 @@
 #ifndef KEPLER
 #define KEPLER
-void propogateOrbitalElements(float semi_major_axis, float eccentricity, float inclination, float ascending_node, float periapsis, float true_anomaly, float time_delta, float *array);
+void propogateOrbitalElements(double semi_major_axis, double eccentricity, double inclination, double ascending_node, double periapsis, double true_anomaly, double time_delta, double *array);
 #endif
