+++ include/moments2currents.h
@@ -1,4 +1,4 @@
 #ifndef MOMENTS2CURRENTS
 #define MOMENTS2CURRENTS
-void moment2current3axis(float* m, float* Imax, float* I_out);
+void moment2current3axis(double* m, double* Imax, double* I_out);
 #endif
