+++ include/photodiode_determination.h
@@ -3,9 +3,9 @@
 
 #include <stdbool.h>
 
-extern const float MAX_READING;
-extern const float PHOTODIODES[18][3];
-extern const float MIN_READING;
-bool get_vec_from_photodiode_readings(float* photodiode_readings, float* estimated_sun_vector);//returns whether we're in the sun or not
+extern const double MAX_READING;
+extern const double PHOTODIODES[18][3];
+extern const double MIN_READING;
+bool get_vec_from_photodiode_readings(double* photodiode_readings, double* estimated_sun_vector);//returns whether we're in the sun or not
 
 #endif
