+++ include/sunvec.h
@@ -1,4 +1,4 @@
 #ifndef SUNVEC
 #define SUNVEC
-void sun_vec(int unix, float* sun);
+void sun_vec(int unix, double* sun);
 #endif
