+++ include/torque2moments.h
@@ -1,4 +1,4 @@
 #ifndef TORQUE2MOMENTS
 #define TORQUE2MOMENTS
-void torque_2_moments(float* B, float* torques, float* moments);
+void torque_2_moments(double* B, double* torques, double* moments);
 #endif
